backend/requirements.txt
# requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
biopython==1.83
numpy==1.26.3
aiofiles==23.2.1